some package reorganisation; some reformatting
